requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
pyyaml
pynvml
numpy==1.23.5
pyworld
scipy==1.10.0
SoundFile==0.12.1
torch
torchaudio
torchcrepe
tqdm
rich
loguru
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
fairseq==0.12.2
librosa==0.9.1
tensorboard
tensorboardX
transformers
edge_tts
langdetect
faiss-cpu
einops
local_attention
audio-separator
